Move proto files to public package.
