--- internal/proto/gen.sh
@@ -1,3 +1,0 @@
-#!/bin/sh
-protoc -I./../../proto/ --go_out=plugins=grpc,import_path=proto:. ./../../proto/*.proto
-
