+++ proto/gen.sh
@@ -1,0 +1,3 @@
+#!/bin/sh
+protoc -I. --go_out=plugins=grpc,import_path=proto:. ./*.proto
+
